autoimport: add progress bar overlay on mobile
